requirements.txt
tensorflow==2.7.0
tensorflow-addons==0.15.0
tensorflow-probability==0.15.0
tensorflow_text
keras==2.7.0
scikit-image
scikit-learn
scipy
six
numpy
PyYAML
absl-py
